Some stuff ig
started proper backend (non-UI) logic for mounting and then accessing the hosting servers. So far, it's just verifying the os platform and connection status of the servers, and not really doing much if they aren't right. Need to remember where to put the await in for the return value of the integrity check.