Bug Fixes
Fixed randomly huge refresh buttons on the home page.

Added a background image as a placeholder and made the window movement bar functionality over the top again.

--Known Issues--
Window can't be resized by dragging; this would present an issue as the contents don't all scale anyway so that might not be so much of a problem right now. Might make UI elements smaller in future so smaller displays can be fit without scaling.